pyproject.toml
[tool.poetry]
name = "fastapi pet"
version = "0.1.0"
description = "pet project for training fastapi skills"
authors = ["PEPPACHAN <PEPPACHAN@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
